# docs/swagger.yaml
basePath: /
definitions:
...
  lib.Order:
    properties:
      cinema_name:
        type: string
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      movie_title:
        type: string
      order_number:
        type: string
      phone_number:
        type: string
      profile_id:
        type: integer
      seats:
        items:
          $ref: '#/definitions/lib.Seat'
        type: array
      show_date:
        type: string
      show_time:
        type: string
      showtime_id:
        type: integer
      status:
        type: string
      total_price:
        type: integer
    type: object
  lib.OrderCreateRequest:
    properties:
      seat_ids:
        items:
          type: integer
        type: array
      showtime_id:
        type: integer
    required:
    - seat_ids
    - showtime_id
    type: object
...
  lib.Payment:
    properties:
      expired_at:
        type: string
      id:
        type: integer
      order_id:
        type: integer
      payment_method:
        type: string
      payment_status:
        type: string
      qr_code:
        type: string
      total_payment:
        type: integer
    type: object
  lib.PaymentRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      payment_method:
        type: string
      phone_number:
        type: string
    required:
    - email
    - full_name
    - payment_method
    - phone_number
    type: object
...
  lib.Response:
    properties:
      message:
        type: string
      result: {}
      status:
        type: integer
    type: object
...
  lib.Seat:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
    type: object
  lib.SeatWithStatus:
    properties:
      id:
        type: integer
      is_occupied:
        type: boolean
      name:
        type: string
      price:
        type: integer
    type: object
...
  lib.UserProfile:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      image:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      point:
        type: integer
    type: object
...
host: localhost:8888
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for Movie Ticket Booking System
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Tikitz API
  version: "1.0"
paths:
...
  /orders:
    post:
      consumes:
      - application/json
      description: 'Step 2: Initialize an order with selected seats'
      parameters:
      - description: Order Details
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/lib.OrderCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                result:
                  $ref: '#/definitions/lib.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - BearerAuth: []
      summary: Create an order
      tags:
      - orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve order summary for the Payment Page
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                result:
                  $ref: '#/definitions/lib.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get order details
      tags:
      - orders
  /orders/{id}/payment:
    post:
      consumes:
      - application/json
      description: 'Step 3: Submit personal info and payment method to complete the
        order'
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payment Details
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/lib.PaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                result:
                  $ref: '#/definitions/lib.Payment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Process payment
      tags:
      - orders
  /orders/{id}/ticket:
    get:
      consumes:
      - application/json
      description: Retrieve the final ticket summary for a specific order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                result:
                  $ref: '#/definitions/lib.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get ticket result
      tags:
      - orders
  /orders/seats:
    get:
      consumes:
      - application/json
      description: Retrieve all seats and their occupancy status for a specific showtime
      parameters:
      - description: Showtime ID
        in: query
        name: showtime_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/lib.SeatWithStatus'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get seats for showtime
      tags:
      - orders
  /users:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all users (admin access required)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/lib.UserProfile'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a user account by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieve a user profile by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                result:
                  $ref: '#/definitions/lib.UserProfile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
  /users/history:
    get:
      consumes:
      - application/json
      description: Retrieve the order history of the currently logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/lib.Order'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - BearerAuth: []
      summary: Get user orders
      tags:
      - users
  /users/profile:
    get:
      consumes:
      - application/json
      description: Retrieve the profile of the currently logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                result:
                  $ref: '#/definitions/lib.UserProfile'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - users
    patch:
      consumes:
      - multipart/form-data
      description: Update the profile of the currently logged-in user
      parameters:
      - description: Full Name
        in: formData
        name: fullName
        type: string
      - description: Email
        in: formData
        name: email
        type: string
      - description: Phone Number
        in: formData
        name: phoneNumber
        type: string
      - description: Profile Image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                result:
                  $ref: '#/definitions/lib.UserProfile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"